ceph-dev-new-setup: Keep more builds
I've seen ceph builds fail because they took so long to claim a build node
that the artifacts created by their ceph-dev-new-setup build was
garbage-collected.

Artifacts for a given build of this job use less than 1G; we should keep more
artifacts around. Metadata for builds of this job are negligible; we should
keep far more sets of metadata around.

Signed-off-by: Zack Cerza <zack@redhat.com>